Penalty function on the composite L2, L2-Square, and L1 penalties

Penalty function on the composite L2, L2-Square, and L1 penalties

penalty(x, lambda, hierarchy, d)

Arguments

x
A vector of two numeric values, in which \(x_1\) represents the prognostic effect, and \(x_2\) for the predictive effect, respectively.

lambda
a vector of three penalty parameters. \(\lambda_1\) and \(\lambda_2\) are L2 and L2-Square (ridge) penalties for \(x\) in a group level, and \(\lambda_3\) is the L1 penalty for \(x_2\), respectively.

hierarchy
a factor value in levels 0, 1, 2, which represent different hierarchical structure in x, respectively. When hierarchy=0, \(\lambda_2\) and \(\lambda_3\) are forced to be zeroes; when hierarchy=1, \(\lambda_2\) is forced to be zero; when hierarchy=2, there is no constraint on \(\lambda\)'s. See smog.

d
indices for overlapped variables in x.

Value

A numeric value of the penalty function based on the composition of L2, L2-Square, and L2 penalties.

References

Ma C, Deng W, Ma S, Liu R, Galinsky K (2019). “Structural modeling using overlapped group penalties for discovering predictive biomarkers for subgroup analysis.” arXiv preprint arXiv:1904.11648.

See also

Examples

penalty(x = rnorm(6,2,1), lambda = c(0.5,0.3,0.1), hierarchy = 0, d = c(1,1,2,2,3,3))

#> [1] 2.656699