Simulate a randomized clinical trial with biomarkers

sim_rct_biomarker is used to simulate clinical trial data with specified treatment, prognostic, and predictive effect sizes.

sim_rct_biomarker(n = 50, p = 100, p_prog = 5, p_pred = 5,
  p_both = 5, v_trt = 0.4, v_prog = 0.2, v_pred = 0.2,
  v_err = 0.2, corr = NULL, family = "gaussian", ...)

Arguments

n
Number of subjects.

p
Number of biomarkers.

p_prog
Number of biomarkers with prognostic effects only.

p_pred
Number of biomarkers with predictive effects only.

p_both
Number of biomarkers with both prognostic and predictive effects

v_trt
Variance of response due to treatment.

v_prog
Variance of response due to prognostic effects.

v_pred
Variance of response due to predictive effects.

v_err
Variance of response due to random noise.

corr
Autocorrelation parameter between biomarkers, default is NULL.

family
The distribution family for response variable, can be gaussian'', or binomial''. Default is ``gaussian''.

...
further arguments passed to or from other methods.

Value

A list containing several variables.

T

Treatment status in 1 or -1 values.

X

Biomarkers.

W

Hadamard product of treatment and biomarkers.

M

Model matrix - binding of T, X, and W.

Y

Response.

Y0

Response without error.

tau

Treatment effect.

beta

Prognostic effects.

gamma

Predictive effects.

theta

All effects corresponding to M.

References

Ma C, Deng W, Ma S, Liu R, Galinsky K (2019). “Structural modeling using overlapped group penalties for discovering predictive biomarkers for subgroup analysis.” arXiv preprint arXiv:1904.11648.

Examples

sim <- sim_rct_biomarker(n = 1e3)
var(as.vector(sim$T * sim$tau))

#> [1] 0.4004004

var(as.vector(sim$X %*% sim$beta))

#> [1] 0.1975435

var(as.vector(sim$W %*% sim$gamma))

#> [1] 0.2190101